+++ src/Components/Home.tsx
@@ -1,0 +1,14 @@
+import { Link } from "react-router-dom";
+
+function Home() {
+    return (
+        <div className="Home">
+            <p><Link to="/Seasons">Seasons</Link></p>
+            <p><Link to="/Seasons/1">First Season</Link></p>
+            <p><Link to="/Types">Types</Link></p>
+            <p><Link to="/Tunes">Tunes</Link></p>
+        </div>
+    );
+}
+
+export default Home;
