+++ src/Components/SeasonDetails.tsx
@@ -1,10 +1,10 @@
 import { useParams } from "react-router-dom";
 
 function SeasonDetails() {
-  let { seasonId } = useParams();
-  return (
-    <div>Season Details {seasonId}</div>
-  );
+    let { seasonId } = useParams();
+    return (
+        <div>Season Details {seasonId}</div>
+    );
 }
 
 export default SeasonDetails;
